+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+### 0.0.8
+- iOS: Add Taptic and Haptic support
+- iOS/AND: vibrate now uses platform namespace i.e. `vibrator.ios::vibrate()`  `vibrator.android::vibrate()`
+
 ### 0.0.7
 - AND: Updated to FreKotlin 1.5.0
 
 ### 0.0.6
-- OSX/iOS: Updated to FreSwift 2.5.0
+- iOS: Updated to FreSwift 2.5.0
 - AND: Updated to FreKotlin 1.4.0
 
 ### 0.0.5